pyproject.toml
[tool.poetry]
name = "pinnjax"
version = "0.1.0"
description = ""
authors = ["ashould01"]

[tool.poetry.dependencies]
python = "^3.10"
matplotlib = "^3.10.0"
jax = {extras = ["cuda11"], version = "^0.5.0"}
optax = "^0.2.4"
tqdm = "^4.67.1"
PyYAML = "^6.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
